requirements.txt
customtkinter>=5.2.2
pypresence>=4.3.0
Pillow>=10.0.0
zeroconf>=0.131.0
requests>=2.31.0
adb-shell>=0.4.4
pycryptodome>=3.19.0
flask>=3.0.0
pywebview>=4.4.1
colorama>=0.4.6
